+++ src/blocks/components/BlockParameter.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import { CodeTemplateBlockParameter } from "../CodeTemplateBlock";
+import { BlockParameterBoolean } from "./BlockParameterBoolean";
+import { makeStyles } from "@fluentui/react-components";
+
+const useStyles = makeStyles({
+    text: {
+        paddingLeft: "10px",
+        paddingRight: "10px",
+    },
+});
+
+export function BlockParameter({
+    parameter,
+    updateValue,
+}: {
+    parameter: CodeTemplateBlockParameter;
+    updateValue: (value: unknown) => void;
+}) {
+    const styles = useStyles();
+    const { name, description, type } = parameter;
+
+    if (type === "boolean") {
+        return (
+            <div className={styles.text}>
+                <BlockParameterBoolean name={name} description={description} updateValue={updateValue} />
+            </div>
+        );
+    }
+    return <div> Unknown Parameter</div>;
+}
